\name{GeneErrorMis-package}
\alias{GeneErrorMis-package}
\alias{GeneErrorMis}
\docType{package}
\title{
  Statistical Tools for Measurement Error and Misclassification in
 Statistical Genetic Studies
}
\description{
  Implementation of insertion strategy for data with mixed measurement error and misclassification in responses. The R package is general to different study type, including known mismeasurement process, available internal validation data and available external validation data.
}
\details{
This package is the toolbox for the jointly modeling the multiple responses in the presence of measurement error and misclassification. The package includes the induced likelihood method, EM algorithm and generalized linear equation to deal with difference scenarios. 
  
}
\author{
Qihuang Zhang and Grace Y. Yi.

Maintainer: Qihuang Zhang <qihuang.zhang@uwaterloo.ca>
}

\keyword{ package, statistical genetics}

