\name{ToyGEEknown}
\docType{data}
\alias{ToyGEEknown}
\title{Toy example data}
\description{
  This data set gives an example data for the illustration of usage of \code{\link{GEEknown}} function. The dataset contains naivedata and the design matrices for the response model (\code{DesignMatrix}), measurement error model (\code{CovMis1}) and misclassification model (\code{CovMis2}). The naive data contain the error-prone responses as well as covariates. This component is mainly used in the naive model. The design matrix of the response model contains the intercept.
}
\usage{data(ToyGEEknown)}
\format{A list of four matrices.}
\keyword{datasets}